climbingTiles: refactor - extract helpers
