+++ src/server/climbing-tiles/refreshClimbingTiles.ts
@@ -1,45 +1,16 @@
-import {
-  GeojsonFeature,
-  OsmResponse,
-  overpassToGeojsons,
-} from './overpass/overpassToGeojsons';
+import { OsmResponse, overpassToGeojsons } from './overpass/overpassToGeojsons';
 import { encodeUrl } from '../../helpers/utils';
 import { fetchJson } from '../../services/fetch';
-import { LineString, LonLat, Point } from '../../services/types';
 import format from 'pg-format';
-import { ClimbingFeaturesRecords, closeClient, getClient } from './db';
+import { closeClient, getClient } from './db';
 import { queryTileStats, updateStats } from './utils';
 import { chunk } from 'lodash';
 import { readFileSync } from 'fs';
-
-const centerGeometry = (feature: GeojsonFeature): GeojsonFeature<Point> => ({
-  ...feature,
-  geometry: {
-    type: 'Point',
-    coordinates: feature.center,
-  },
-});
-
-const firstPointGeometry = (
-  feature: GeojsonFeature<LineString>,
-): GeojsonFeature<Point> => ({
-  ...feature,
-  geometry: {
-    type: 'Point',
-    coordinates: feature.geometry.coordinates[0],
-  },
-});
-
-const prepareGeojson = (
-  type: string,
-  { id, geometry, properties }: GeojsonFeature,
-) =>
-  JSON.stringify({
-    type: 'Feature',
-    id,
-    geometry,
-    properties: { ...properties, type },
-  });
+import {
+  buildLogFactory,
+  centerGeometry,
+  recordsFactory,
+} from './refreshClimbingTilesHelpers';
 
 const fetchFromOverpass = async () => {
   if (process.env.NODE_ENV === 'development') {
@@ -65,43 +36,6 @@ const fetchFromOverpass = async () => {
   }
 
   return data;
-};
-
-const removeDiacritics = (str: string) =>
-  str?.normalize('NFD').replace(/[\u0300-\u036f]/g, '');
-
-const recordsFactory = () => {
-  const records: ClimbingFeaturesRecords = [];
-  const addRecordRaw = (
-    type: string,
-    coordinates: LonLat,
-    feature: GeojsonFeature,
-  ) => {
-    const lon = coordinates[0];
-    const lat = coordinates[1];
-    return records.push({
-      type,
-      osmType: feature.osmMeta.type,
-      osmId: feature.osmMeta.id,
-      name: feature.tags.name,
-      nameRaw: removeDiacritics(feature.tags.name),
-      count: feature.properties.osmappRouteCount || 0,
-      lon,
-      lat,
-      geojson: prepareGeojson(type, feature),
-    });
-  };
-
-  const addRecord = (type: string, feature: GeojsonFeature<Point>) => {
-    addRecordRaw(type, feature.geometry.coordinates, feature);
-  };
-
-  const addRecordWithLine = (type: string, way: GeojsonFeature<LineString>) => {
-    addRecord(type, firstPointGeometry(way));
-    addRecordRaw(type, way.center, way);
-  };
-
-  return { records, addRecord, addRecordWithLine };
 };
 
 // eslint-disable-next-line max-lines-per-function
@@ -205,16 +139,6 @@ const getNewRecords = (data: OsmResponse) => {
   }
 
   return records;
-};
-
-const buildLogFactory = () => {
-  const buildLog: string[] = [];
-  const log = (message: string) => {
-    buildLog.push(message);
-    console.log(message); //eslint-disable-line no-console
-  };
-  log('Starting...');
-  return { getBuildLog: () => buildLog.join('\n'), log };
 };
 
 export const refreshClimbingTiles = async () => {
